Update chapter number reference in ChapterListBox
Changed the chapter number reference from 'getChapterNumber()' to 'getIndex()' in ChapterListBox.java. This has been an Error previously. On normal Manga this wouldn't do anything different, but on Manga with a Prologue, usually named chapter 0, it would result in an Exception being thrown. The Index always is above 0 and is also what's being used to retrieve the Images from the Tachidesk server.