Add library update button and related functionalities
Added a new button in the library UI for refreshing the library data with a new library update functionality. Introduced a new service, LibUpdateService which is responsible for fetching the updates. This was necessary to provide users with an intuitive way to update their library content, ensuring they always have the latest manga data available. Alongside, CSS changes are also implemented to provide a nice UI.