Add read/unread status change functionality to chapter list
Updated the ChapterRenderer and ChapterListBox to support changing the read/unread status of manga chapters. Now users can mark chapters as read or unread directly from the list, improving user interaction and experience. A new event 'ChapterReadStatusChangeEvent' has been added to fire when the read status of a chapter changes. The MangaService has been updated with methods to change the read status in the backend.

Note: You will only get redirected to a new chapter if you press on the actual blue part of the list item, as otherwise you wouldn't be able to interact with the buttons