Add resume reading feature and refactor Manga class
This commit adds a 'Resume Reading' button to the Manga view, this feature enables the user to continue from the last chapter they left off. Additionally, the Manga class has been refactored, the getters were removed and the Lombok @Getter annotation is applied to the class to handle this. Class RouteUtils has been added to handle routing to the reading view. The changes also include refactoring services' classes and methods in relation to needed features for the Resume Feature.