Refactor MangaView - remove RestTemplate dependency
Assessed the MangaView class and concluded that the RestTemplate dependency was obsolete. This is due to the fact that our client-side is not utilizing the RestTemplate to make its HTTP requests at the moment. Subsequently, the pertaining import and method parameter were eliminated to maintain clean code. Furthermore, the remaining HTTP requests are dealt with via Spring Boot from other classes such as MangaService, SettingsService, and AniListAPIService.