Improve MangaReader equality check and Chapter class
Added an equality check in MangaReader to prevent unnecessary mutations and updates when the old and new values are the same. This is achieved using the Objects.equals() method. This check adds an extra layer of protection against possible redundant operations. Also, an "EqualsAndHashCode" annotation is added to the Chapter class to handle potential equality checks between Chapter objects, as before the equality check didn't work correctly