Remove StartingService, add browser starting in TrayHandler
Removed the unnecessary StartingService that was starting a browser upon application event. This responsibility was transferred to TrayHandler where upon startup, a message is displayed, guiding the user how to access the application. A click event is also added to the tray icon, to launch a browser window targeting the application URL. This change improves the user experience and application initialization reliability.