Add CategoryService to handle category creation
Introduced a new CategoryService to take responsibility of category creation instead of the previous implementation where the task was scattered across CategoryDialog and CategoryUtils. This change centralizes the creation logic, thus increasing maintainability and adaptability for future extensions. The CategoryService utilizes a CategoryClient interface to implement the category creation function. The service is then injected and used in CategoryDialog and RootView classes. All CategoryUtils functions will be implemented in CategoryService/CategoryClient in the next commits