Updated SettingsService.java to ensure server data directory is OS specific.
The change is introduced to make the application more robust on different operating systems. Previously, SettingsService used a static path to store server data, which is not reliable if the application is deployed on different OS. Now, the TachideskMaintainer component is used to determine the project's directory dynamically and it optimizes the server data storage path. As a result, server data directory is now OS specific.