Add native installers and server startup checks
Introduced JPackage plugin to create native installers and added additional checks for server startup. The changes enable the production of platform-specific installers - making application setup more user-friendly. Furthermore, additional checks were incorporated in the server start-up process to ensure smooth operation, this includes checking for the existence of a Java installation. If not present, the user is directed to appropriate help resources.

Note: This only adds support for general native installers and adds tasks for Windows installers. Linux and Mac will come in the future