Refactor RootView
This commit makes several changes for better code maintenance and user-experience. Firstly, in CategoryService.java, the error log level for getting categories has been lowered to debug. This change is important to reduce unnecessary noise in the error logs as this is not critical and in fact expected, when redirecting to the ServerStartView.

Significant refactoring was done in RootView.java to enhance the readability and maintenance. The view initialization code in the constructor has been moved into the new method "beforeEnter". By leveraging the BeforeEnterObserver interface, we make sure that the view is ready before the user enters it, while avoiding returning prematurely out of the constructor. This also seems to help with Performance when creating the View. Creating the tabSheet buttons are now handled by a separate method getTabSheetButtons() which makes the code cleaner. Also, the SettingsService is now held at an instance level.