Add zoom feature to manga reader and update FlowSwiper version
Added a zoom functionality to the manga reader and updated the FlowSwiper package to version 1.3.0 in the build.gradle file. The zoom feature was implemented by adding the zoom state to the swiper configuration in the MangaReader class, and tie it to scale changes through event listeners. This provides a better reading experience to the user, allowing them to control the view of the manga pages. Changes in gradle dependencies have been made in order to support this new feature.