Update Authorization flow to Implicit Grant for AniList API.
This commit transitions the AniList API's OAuth2 authentication from "Authorization Code" to "Implicit Grant". This change enables us to not include a Client Secret anymore which drastically simplifies the login process, as well as the security of the OAuth Process. OAuthData and TrackerTokens were also updated to store the extracted token in a new format.