Update TrackingDialog for AniList button
Modify tracking dialog check for AniList. Now, it checks not only if tracker has a AniList Id associated with it, but also if the user is authenticated with an AniList token. This makes sure that the user can authenticate themselves again, if the Token was removed. This update enhances user experience by removing potential "AniList" errors in the TrackingDialog flow