Refactor project directory retrieval into a utility class
The project directory retrieval, previously located in TachideskMaintainer.java, was moved to a newly created utility class PathUtils.java for better code organization and reuse. As a result internal methods in TachideskMaintainer.java and a set of dependencies in services classes, SettingsService.java and TrackingDataService.java, were updated accordingly to utilize this new utility method. Reason for these changes is to unify the retrieval logic of the project directory into a single place for easier maintainability, as well as avoiding circular dependencies in the future.