Add URL change event and listeners
This commit adds a URL change event and adds listeners to different services to handle this event. This was done to ensure that whenever the stored URL in the settings is changed, all respective services, including `TachideskMaintainer` and `TachideskStarter`, are notified and can take appropriate actions. The setting view now publishes the URL change event whenever a new URL is set in the system settings. The logic was added inside the maintained and starter to restart the services whenever URL changes occur. Also, a method was added to retrieve default settings to make possible the comparison with the new values.