Refactor search executor to be within try block
Moved executor creation into try-with-resources block in SearchView.java, ensuring threads are properly shut down even if an exception is thrown. This handles potential for resource leak due to unclosed executor. This is being added here as I found out about it while managing this branch and didn't want to forget about it.