Add reader settings dialog and update settings logic
Added a new 'ReaderSettingsDialog' file and logic for the manga reader to configure and save reader settings either as default or specific to a manga. These settings include direction of reading which can be Right-to-Left or Left-to-Right. New files 'ReaderSettings' and 'ReaderDirection' hold the settings data. 'Settings.java' was updated to include logic managing reader specific settings. User interface updates to support these changes were also made in 'manga-reader.css', 'reader-settings-dialog.css', and 'manga-reader.java'.