Update reader settings to use defaults
Updated ReaderSettingsDialog to use default settings when specific manga settings aren't available. The 'Save as Default' button now works with the default settings rather than the Settings provided by Settings#getReaderSettings. Added a getter for defaultReaderSettings and a method to check if mangaReaderSettings are available. The latter will be used to only send update events to the UI if there's no specific Manga settings, when the default settings are being changed.