Add AniList integration to Manga search feature
This commit introduces a new AniList view and related classes, which enable users to search for Manga from their AniList Manga List. A new button was added to the SearchView that allows navigation to the new AniList view for importing Manga. This feature facilitates an easier and less time-consuming way for users to migrate and track Manga from their AniList account.