Remove unnecessary install check from ExtensionItem
This commit removes an unnecessary check for whether an extension is installed in the ExtensionItem.java file. Before this change, installed extensions would always display as 'Installed', regardless of whether they had an update available or were obsolete. This update ensures that the correct status is displayed for each extension.

Also made buttons change their status instantly instead of having to reload the page for changes to appear