Implement download tracking with Websockets for efficiency
This commit introduces two major changes to improve the efficiency of download tracking. Firstly, the GraphQLs `trackDownloads` subscription is now used to track downloads in real-time through Websockets. This has been implemented in the `ChapterRenderer.java` and `DownloadClient.java` classes using the Flux API, instead of frequently querying the server about the download status. A new WebSocketGraphQlClient has been introduced for use with the GraphQL subscription in `WebClientService.java`. This process now keeps track of the downloads via events, eliminating the need to continuously ask the server about the status of chapter downloads. This method is much more efficient and provides real-time updates about the download progress. Furthermore, the `initWebSocketGraphQlClient()` method was created in the `WebClientService` class to initialize the WebSocket client. Additionally, the `destroy()` method has been created in the same class to stop the WebSocket client from blocking the shutdown process. As a result, the UIs processing need is greatly reduced, making the system overall more efficient.