Add support for dev profile and directory
Introduced development profiles to the application ensuring that during development, all work is done in a new development directory instead of the default project directory. This is done by updating the TrackingDataService, TachideskMaintainer, SettingsService, and PathUtils, establishing a new file, and making adjustments in build.gradle. This helps in preventing changes from affecting normal data on a devs computer. For example overriding the private Library of the Dev