Refine server connection check route
The server connection verification route was updated from "/api/v1/meta" to "/api/v1/settings/about". The change was required as the previous endpoint relied on SQL data which is not guaranteed to be available at the moment of the check, causing potential connection testing failure. Exception handling was also refined to specifically catch FeignExceptions.