Improve logic for choosing the next chapter
This commit changes the logic of the "Resume" button in 'MangaView'. Previously, if the last chapter read was null, the next chapter was set as the last in the chapter list. Now, the code checks the chapters in reverse order to find the next unread one. If no such chapters are found, the last chapter is used instead